+++ packages/components/psammead-bulletin/src/index.test.jsx
@@ -14,18 +14,30 @@ const BulletinComponent = ({
   withSummary = true,
   dir = 'ltr',
   lang = null,
+  ariaId,
 }) => {
   const summaryText = 'This is the summary text';
   const headlineText = 'This is the headline';
   const ctaLink = 'https://bbc.co.uk';
+  const imageSizes = [300, 450, 600, 1024];
+  const imageSrc =
+    'https://ichef.bbci.co.uk/news/[WIDTH]/cpsprodpb/11897/production/_106613817_999_al_.jpg';
 
   const playCtaText = isLive ? `${ctaText} Live` : ctaText;
   const offScreenText = isLive ? `${ctaText} LIVE` : ctaText;
 
   const image = (
     <Image
-      src="https://ichef.bbci.co.uk/news/660/cpsprodpb/11897/production/_106613817_999_al_.jpg"
+      src={imageSrc}
       alt="Iron man"
+      srcset={imageSizes
+        .map(size => `${imageSrc.replace('[WIDTH]', size)}.webp ${size}w`)
+        .join(', ')}
+      fallbackSrcset={imageSizes
+        .map(size => `${imageSrc.replace('[WIDTH]', size)} ${size}w`)
+        .join(', ')}
+      primaryMimeType="image/webp"
+      fallbackMimeType="image/jpeg"
     />
   );
   return (
@@ -42,6 +54,7 @@ const BulletinComponent = ({
       offScreenText={offScreenText}
       dir={dir}
       lang={lang}
+      ariaId={ariaId}
     />
   );
 };
@@ -54,6 +67,7 @@ describe('Bulletin', () => {
       service="news"
       mediaType="audio"
       ctaText="Listen"
+      ariaId="https://bbc.co.uk"
     />,
   );
 
@@ -66,6 +80,7 @@ describe('Bulletin', () => {
       ctaText="Listen"
       dir="rtl"
       lang="en-GB"
+      ariaId="https://bbc.co.uk"
     />,
   );
 
@@ -76,6 +91,7 @@ describe('Bulletin', () => {
       service="news"
       mediaType="video"
       ctaText="Watch"
+      ariaId="https://bbc.co.uk"
     />,
   );
 
@@ -86,6 +102,7 @@ describe('Bulletin', () => {
       service="news"
       mediaType="audio"
       ctaText="Listen"
+      ariaId="https://bbc.co.uk"
       isLive
     />,
   );
@@ -97,6 +114,7 @@ describe('Bulletin', () => {
       service="news"
       mediaType="video"
       ctaText="Watch"
+      ariaId="https://bbc.co.uk"
       isLive
     />,
   );
@@ -108,7 +126,19 @@ describe('Bulletin', () => {
       service="news"
       mediaType="audio"
       ctaText="Listen"
+      ariaId="https://bbc.co.uk"
       withSummary={false}
+    />,
+  );
+
+  // the below test is a temporary test for the a11y nested span's bug experienced in TalkBack, refer to the following issue: https://github.com/bbc/simorgh/issues/9652
+  shouldMatchSnapshot(
+    'should render radio bulletin without ariaId',
+    <BulletinComponent
+      script={latin}
+      service="news"
+      mediaType="audio"
+      ctaText="Listen"
     />,
   );
 });