+++ packages/components/psammead-image/src/index.amp.jsx
@@ -3,12 +3,26 @@ import omit from 'ramda/src/omit';
 import { number, string } from 'prop-types';
 
 // Prevents component outputting invalid HTML when styled with emotion
-const omitInvalidProps = omit(['classname']);
+const omitInvalidProps = omit([
+  'classname',
+  'primaryMimeType',
+  'fallbackMimeType',
+]);
 
 const AmpImg = props => {
-  const { srcset, ...otherProps } = props;
+  const { srcset, fallbackSrcset, ...otherProps } = props;
 
-  return <amp-img srcSet={srcset} {...omitInvalidProps(otherProps)} />;
+  return (
+    <amp-img srcSet={srcset} {...omitInvalidProps(otherProps)}>
+      {fallbackSrcset && (
+        <amp-img
+          fallback=""
+          srcSet={fallbackSrcset}
+          {...omitInvalidProps(otherProps)}
+        />
+      )}
+    </amp-img>
+  );
 };
 
 AmpImg.propTypes = {
@@ -19,6 +33,7 @@ AmpImg.propTypes = {
   sizes: string,
   src: string.isRequired,
   srcset: string,
+  fallbackSrcset: string,
   width: number.isRequired,
 };
 
@@ -26,6 +41,7 @@ AmpImg.defaultProps = {
   attribution: '',
   sizes: null,
   srcset: null,
+  fallbackSrcset: null,
 };
 
 export default AmpImg;