+++ packages/components/psammead-image/src/index.stories.jsx
@@ -1,21 +1,15 @@
-import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { Img } from '.';
-import { stories, getProps } from './testHelpers/stories';
-import { landscape } from './testHelpers/fixtureData';
-import notes from '../README.md';
+import { stories } from './testHelpers/stories';
 
 const type = 'Img';
 
-stories(
-  Img,
-  'Components/Images/Image - Img',
-  false,
-  {},
-  withKnobs,
+stories({
+  Component: Img,
+  title: 'Components/Images/Image - Img',
+  includeHeight: false,
+  additionalProps: {},
+  styleDecorator: withKnobs,
   type,
-).add(
-  'image without width',
-  () => <Img {...getProps(landscape, false, type)} width={null} />,
-  { notes },
-);
+  isCanonical: true,
+});