+++ packages/components/psammead-navigation/src/DropdownNavigation/index.jsx
@@ -129,20 +129,33 @@ export const DropdownLi = ({
   service,
   url,
   dir,
-}) => (
-  <StyledDropdownLi role="listitem">
-    <StyledDropdownLink script={script} service={service} href={url}>
-      {active && currentPageText ? (
-        <StyledCurrentLink dir={dir}>
-          <VisuallyHiddenText>{currentPageText}, </VisuallyHiddenText>
-          {children}
-        </StyledCurrentLink>
-      ) : (
-        children
-      )}
-    </StyledDropdownLink>
-  </StyledDropdownLi>
-);
+}) => {
+  const ariaId = `dropdownNavigation-${children
+    .replace(/\s+/g, '-')
+    .toLowerCase()}`;
+  return (
+    // aria-labelledby is a temporary fix for the a11y nested span's bug experienced in TalkBack, refer to the following issue: https://github.com/bbc/simorgh/issues/9652
+    <StyledDropdownLi role="listitem">
+      <StyledDropdownLink
+        script={script}
+        service={service}
+        href={url}
+        aria-labelledby={ariaId}
+      >
+        {active && currentPageText ? (
+          // ID is a temporary fix for the a11y nested span's bug experienced in TalkBack, refer to the following issue: https://github.com/bbc/simorgh/issues/9652
+          <StyledCurrentLink dir={dir} id={ariaId}>
+            <VisuallyHiddenText>{`${currentPageText}, `}</VisuallyHiddenText>
+            {children}
+          </StyledCurrentLink>
+        ) : (
+          // ID is a temporary fix for the a11y nested span's bug experienced in TalkBack, refer to the following issue: https://github.com/bbc/simorgh/issues/9652
+          <span id={ariaId}>{children}</span>
+        )}
+      </StyledDropdownLink>
+    </StyledDropdownLi>
+  );
+};
 
 DropdownLi.propTypes = {
   children: string.isRequired,