+++ packages/utilities/psammead-styles/src/colours.js
@@ -40,3 +40,4 @@ export const C_SPORT_SILVER = '#DBDBDB';
 export const C_SPORT_MIST = '#F7F7F5';
 export const C_BLACK = '#000000';
 export const C_GREY_6 = '#545658';
+export const C_GREY_8 = '#202224';