+++ packages/components/psammead-media-player/CHANGELOG.md
@@ -3,6 +3,11 @@
 <!-- prettier-ignore -->
 | Version       | Description                                                                                                                  |
 | ------------- | ---------------------------------------------------------------------------------------------------------------------------- |
+| 6.0.3 | [PR#4609](https://github.com/bbc/psammead/pull/4609) Bump from psammead-styles |
+| 6.0.2| [PR#4608](https://github.com/bbc/psammead/pull/4608) Bump dependencies|
+| 6.0.1| [PR#4607](https://github.com/bbc/psammead/pull/4607) Bump dependencies|
+| 6.0.0| [PR#4606](https://github.com/bbc/psammead/pull/4606) Adds support for WebP |
+| 5.1.13 | [PR#4601](https://github.com/bbc/psammead/pull/4601) Bumps dependencies |
 | 5.1.13 | [PR#4574](https://github.com/bbc/psammead/pull/4574) Bumps psammead-styles transitive packages |
 | 5.1.12 | [PR#4568](https://github.com/bbc/psammead/pull/4568) Bump from psammead-styles |
 | 5.1.11 | [PR#4565](https://github.com/bbc/psammead/pull/4565) Bump from psammead-styles |