+++ packages/components/psammead-story-promo-list/CHANGELOG.md
@@ -3,6 +3,9 @@
 <!-- prettier-ignore -->
 | Version | Description |
 | ------- | ----------- |
+| 6.0.31 | [PR#4609](https://github.com/bbc/psammead/pull/4609) Bump from psammead-styles |
+| 6.0.30 | [PR#4608](https://github.com/bbc/psammead/pull/4608) Bumps psammead-styles |
+| 6.0.29 | [PR#4606](https://github.com/bbc/psammead/pull/4606) Bumps psammead-styles |
 | 6.0.28 | [PR#4574](https://github.com/bbc/psammead/pull/4574) Bumps psammead-styles |
 | 6.0.27 | [PR#4568](https://github.com/bbc/psammead/pull/4568) Bump dependencies |
 | 6.0.26 | [PR#4565](https://github.com/bbc/psammead/pull/4565) Bump from psammead-styles |