+++ packages/components/psammead-useful-links/CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 3.0.30 | [PR#4609](https://github.com/bbc/psammead/pull/4609) Bump from psammead-styles |
 | 3.0.29 | [PR#4574](https://github.com/bbc/psammead/pull/4574) Bumps psammead-styles |
 | 3.0.28 | [PR#4568](https://github.com/bbc/psammead/pull/4568) Bump dependencies |
 | 3.0.27 | [PR#4565](https://github.com/bbc/psammead/pull/4565) Bump from psammead-styles |